Contrato OK

+++ backend/-1-Domain/Entities/contract.cs
@@ -20,7 +20,7 @@ namespace ERP.Domain.Entities
 		public string Type { get; set; }
 
 		[Column("contract_value")]
-		public long Value { get; set; }
+		public decimal Value { get; set; }
 
 		[Column("contract_is_payroll")]
 		public bool IsPayroll { get; set; }
@@ -72,7 +72,7 @@ namespace ERP.Domain.Entities
 		public Contract(
 			long Param_EmployeeId, 
 			string Param_Type, 
-			long Param_Value, 
+			decimal Param_Value, 
 			bool Param_IsPayroll, 
 			bool Param_HasInss, 
 			bool Param_HasIrrf,
