+++ backend/-2-Application/DTOs/ContractCostCenter/ContractCostCenterOutputDTO.cs
@@ -5,7 +5,7 @@ namespace ERP.Application.DTOs
         public long ContractCostCenterId { get; set; }
         public long ContractId { get; set; }
         public long CostCenterId { get; set; }
-        public long Percentage { get; set; }
+        public decimal Percentage { get; set; }
         public long CriadoPor { get; set; }
         public long? AtualizadoPor { get; set; }
         public DateTime CriadoEm { get; set; }
